ramdisk:stweak: add more & better options to init process fixer

+++ WORKING-DIR/ramdisk/ramdisk/gabriel/ext/post-init.sh
@@ -339,10 +339,20 @@ if [ "$stweaks_init_proc_fixer" == "yes" ]; then
 
 	if [ "$($BB top -n 10 -d 1 | grep init | wc -l)" -gt "5" ] &&
 	   [ "$(cat /data/init_proc_fixer)" -ne "1" ];then
-		$BB cp /system/bin/dpmd /system/bin/dpmd.bak;
-		$BB cp /vendor/bin/msm_irqbalance /vendor/bin/msm_irqbalance.bak;
-		$BB sed -i '1d' /system/bin/dpmd;
-		$BB sed -i '1d' /vendor/bin/msm_irqbalance;
+		if [ ! -e /system/bin/dpmd.bak ]; then
+			$BB cp /system/bin/dpmd /system/bin/dpmd.bak;
+			$BB sed -i '1d' /system/bin/dpmd;
+		else
+			$BB sed -i '1d' /system/bin/dpmd;
+		fi;
+
+		if [ ! -e /vendor/bin/msm_irqbalance.bak ]; then
+			$BB cp /vendor/bin/msm_irqbalance /vendor/bin/msm_irqbalance.bak;
+			$BB sed -i '1d' /vendor/bin/msm_irqbalance;
+		else
+			$BB sed -i '1d' /vendor/bin/msm_irqbalance;
+		fi;
+
 		echo 1 > /data/init_proc_fixer;
 	fi;
 fi;
